оптимизипрован проект

+++ src/pid.cpp
@@ -3,34 +3,35 @@
 
 void pid::start(double set_temp)
 {
-double integral=0,prevTemp=0;
-temp_delta=temp_int-set_temp;
-sprintf(buffer, ">temp_delta:%-8.2f\n",temp_delta);
-uart_1.uart_tx_data(buffer);
-DWT_CONTROL &= ~DWT_CTRL_CYCCNTENA_Msk;
-dt=DWT_CYCCNT/1000;
-sprintf(buffer, ">dt:%d\n",dt);
-uart_1.uart_tx_data(buffer);
-integral+=temp_delta*dt*ki;
-sprintf(buffer, ">PID_I:%-8.2f\n",integral*ki);
- uart_1.uart_tx_data(buffer);
-sprintf(buffer, ">set:%f\n",set_temp);
- uart_1.uart_tx_data(buffer);
+    double integral=0,prevTemp=0;
+    temp_delta=temp_int-set_temp;
+    sprintf(buffer, ">temp_delta:%-8.2f\n",temp_delta);
+    uart_1.uart_tx_data(buffer);
+    DWT_CONTROL &= ~DWT_CTRL_CYCCNTENA_Msk;
+    dt=DWT_CYCCNT/1000;
+    sprintf(buffer, ">dt:%d\n",static_cast<int>(dt));
+    uart_1.uart_tx_data(buffer);
+    integral+=temp_delta*dt*ki;
+    sprintf(buffer, ">PID_I:%-8.2f\n",integral*ki);
+    uart_1.uart_tx_data(buffer);
+    sprintf(buffer, ">set:%f\n",set_temp);
+    uart_1.uart_tx_data(buffer);
 
-double D=(temp_delta-prevTemp)/dt;
-sprintf(buffer, ">PID_D:%f\n",D);
-uart_1.uart_tx_data(buffer);
-prevTemp=temp_delta;
-pwm=temp_delta*kp+integral*ki+D*kd;
-pwm+=100;
-if(pwm>reg_max)
- pwm =reg_max;//-pwm;
- if(pwm<reg_min)
- pwm =reg_min;//-pwm;
-  //TIM3->CCR3 =(uint16_t)pwm;
- TIM3->CCR3 =500;
- sprintf(buffer, ">PWM:%-8.2f\n",pwm);
- uart_1.uart_tx_data(buffer);
+    double D=(temp_delta-prevTemp)/dt;
+    sprintf(buffer, ">PID_D:%f\n",D);
+    uart_1.uart_tx_data(buffer);
+    prevTemp=temp_delta;
+    pwm=temp_delta*kp+integral*ki+D*kd;
+    pwm+=100;
+
+    if(pwm>reg_max) 
+        pwm =reg_max;//-pwm;
+    if(pwm<reg_min) 
+        pwm =reg_min;//-pwm;
+    //TIM3->CCR3 =(uint16_t)pwm;
+    TIM3->CCR3 =500;
+    sprintf(buffer, ">PWM:%-8.2f\n",pwm);
+    uart_1.uart_tx_data(buffer);
  }
 
 // double pid::calc_PID(double setpoint,double kp,double ki,double kd)
